Updated some of the gui code
