+++ src/main/java/de/thowl/automomousinstantdocumentsystem/control/DatabaseScene.java
@@ -1,0 +1,5 @@
+package de.thowl.automomousinstantdocumentsystem.control;
+
+public class DatabaseScene extends MasterController {
+
+}
